Service: Move tests to the right package (#2469)
